ai/veo-games/requirements.txt
# AI Pipeline Automation Dependencies
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0